migrate: move locationsListing to products directory
- Migrated locationsListing product files to src/products/locationsListing/
- Moved and renamed index.tsx -> template.tsx
- Moved and renamed index.test.tsx -> template.test.tsx
- Moved query.ts from src/data/queries/ to src/products/locationsListing/
- Moved test file and renamed locationsListing.test.tsx -> query.test.ts
- Moved formatted type from src/types/formatted/ to src/products/locationsListing/formatted-type.ts
- Moved mock file from src/mocks/ to src/products/locationsListing/mock.json
- Updated all import statements across the codebase
- Fixed import paths in migrated files (formatted-type.ts, query.ts, template.test.tsx, query.test.ts)
- All TypeScript errors resolved
- All tests passing (9 query tests + 17 template tests)
- Updated migration guide with completed status
- Added CODEOWNERS entry for new product directory
- Verified dev server runs and locationsListing pages load correctly (Pittsburgh locations confirmed)
