+++ src/data/queries/index.ts
@@ -20,7 +20,7 @@ import * as FeaturedContent from '@/data/queries/featuredContent'
 import * as HeaderFooter from '@/data/queries/headerFooter'
 import * as HealthCareLocalFacility from '@/products/healthCareLocalFacility/query'
 import * as LinkTeaser from '@/data/queries/linkTeaser'
-import * as LocationsListing from '@/data/queries/locationsListing'
+import * as LocationsListing from '@/products/locationsListing/query'
 import * as MediaDocument from '@/data/queries/mediaDocument'
 import * as MediaImage from '@/data/queries/mediaImage'
 import * as MediaVideo from '@/data/queries/mediaVideo'
