+++ src/pages/[[...slug]].tsx
@@ -44,7 +44,7 @@ const isExport = process.env.BUILD_OPTION === 'static'
 // Types
 import { Event as FormattedEvent } from '@/products/event/formatted-type'
 import { EventListing as FormattedEventListing } from '@/products/eventListing/formatted-type'
-import { LocationsListing as FormattedLocationsListing } from '@/types/formatted/locationsListing'
+import { LocationsListing as FormattedLocationsListing } from '@/products/locationsListing/formatted-type'
 import { NewsStory as FormattedNewsStory } from '@/products/newsStory/formatted-type'
 import { PressRelease as FormattedPressRelease } from '@/products/pressRelease/formatted-type'
 import { PressReleaseListing as FormattedPressReleaseListing } from '@/products/pressReleaseListing/formatted-type'
@@ -64,7 +64,7 @@ import HTMLComment from '@/templates/common/util/HTMLComment'
 import { Event } from '@/products/event/template'
 import { EventListing } from '@/products/eventListing/template'
 import { LayoutProps } from '@/templates/layouts/wrapper'
-import { LocationsListing } from '@/templates/layouts/locationsListing'
+import { LocationsListing } from '@/products/locationsListing/template'
 import { Meta } from '@/templates/common/meta'
 import { NewsStory } from '@/products/newsStory/template'
 import { PressRelease } from '@/products/pressRelease/template'
